requirements.txt
openai
SpeechRecognition
TK
PyAudio
pyttsx3
pillow
